Résolution des problèmes de convergence dû à la précision.
Pour résoudre le problème de convergence la dernière valeur n'est plus
égale à 1-(les autres) mais à la valeur indiquée dans le graphe, cela a
pour conséquence de ne plus toujours converger à exactement 1. La
précision a aussi été augmentée en passant de double à long double et en
ajoutant l'option -mfpmath=387 qui indique au processeur d'utiliser la
précision maximale.

+++ main.c
@@ -16,7 +16,7 @@ int main(int argc, char* argv[]){
 
 	Element** index = NULL;
 	uint8_t* emptyLines = NULL;
-	double* delta = NULL;
+	long double* delta = NULL;
 
 	if(readFile(argv[1], &index, &n, &emptyLines, &delta) == -1){
 		fprintf(stderr, "Fichier non conforme\n");
